feat: refactor sf-store to be _internal
