--- packages/vue/src/components/molecules/SfStore/SfStore.vue
@@ -1,5 +1,0 @@
-<script src="./SfStore.js"></script>
-<template lang="html" src="./SfStore.html"></template>
-<style lang="scss">
-@import "~@storefrontui/shared/styles/components/SfStore.scss";
-</style>
