+++ packages/vue/src/components/organisms/SfStoreLocator/SfSToreLocator.stories.js
@@ -1,9 +1,14 @@
 // /* eslint-disable import/no-extraneous-dependencies */
 import { storiesOf } from "@storybook/vue";
-import { withKnobs, object, number, select } from "@storybook/addon-knobs";
+import {
+  withKnobs,
+  object,
+  number,
+  select,
+  text
+} from "@storybook/addon-knobs";
 import { generateStorybookTable } from "@/helpers";
 import SfStoreLocator from "./SfStoreLocator.vue";
-import SfStore from "../../molecules/SfStore/SfStore.vue";
 
 const stores = [
   {
@@ -59,6 +64,40 @@ const scssTableConfig = {
       "$sf-store-locator__map-min-height-mobile",
       "26.875rem",
       "Store map mobile min height"
+    ],
+    ["$sf-store__media-width", "5.125rem", "Store card media width"],
+    ["$sf-store__media-height", "7rem", "Store card media height"],
+    ["$sf-store__media-margin", "0 1.25rem 0 0", "Store card media margin"],
+    [
+      "$sf-store__item-font-size",
+      "$font-size-extra-small-desktop",
+      "Store card font size"
+    ],
+    [
+      "$sf-store__name-font-size",
+      "$font-size-regular-desktop",
+      "Store name font size"
+    ],
+    [
+      "$sf-store__name-font-weight",
+      "$h1-font-weight-desktop",
+      "Store name font weight"
+    ],
+    ["$sf-store__color-primary", "$c-dark-primary", "Store card primary color"],
+    [
+      "$sf-store__color-secondary",
+      "$c-gray-secondary",
+      "Store card secondary color"
+    ],
+    [
+      "$sf-store__font-family-primary",
+      "$body-font-family-primary",
+      "Store card primary font family"
+    ],
+    [
+      "$sf-store__font-family-secondary",
+      "$body-font-family-secondary",
+      "Store card secondary font family"
     ]
   ]
 };
@@ -129,7 +168,7 @@ storiesOf("Organisms|StoreLocator", module)
           default: object("markerOptions", {})
         }
       },
-      components: { SfStoreLocator, SfStore },
+      components: { SfStoreLocator },
       template: `
       <sf-store-locator :center="center" :zoom="zoom" :max-zoom="maxZoom" :fly-to-store-zoom="flyToStoreZoom" :tile-server-url="tileServerUrl" >
         <sf-store v-for="(store, index) in stores" :key="index" v-bind="store" />
@@ -143,7 +182,49 @@ storiesOf("Organisms|StoreLocator", module)
           center the map on it, distance is calculated based on the user position.
           Even if Leaflet and Vue2Leaflet do not support SSR this component defer the loading of the library to the mounted lifecycle event
           <h2> Usage </h2>
-          <pre><code>import { SfStoreLocator, SfStore } from "@storefrontui/vue"</code></pre>
+          <pre><code>import { SfStoreLocator } from "@storefrontui/vue"</code></pre>
+          ${generateStorybookTable(scssTableConfig, "SCSS variables")}
+          `
+      }
+    }
+  )
+  .add(
+    "SfStore",
+    () => ({
+      props: {
+        latlng: {
+          default: object("latlng", [47.5104387, 19.0444433])
+        },
+        name: {
+          default: text("name", "My Store")
+        },
+        address: {
+          default: text("address", "Store address 100228 NY")
+        },
+        phone: {
+          default: text("phone", "(70) 881 9463")
+        },
+        email: {
+          default: text("email", "some@email.com")
+        },
+        picture: {
+          default: text("picture", "https://picsum.photos/200/300")
+        }
+      },
+      components: { SfStoreLocator },
+      template: `
+      <sf-store-locator :center="latlng">
+        <sf-store :name="name" :picture="picture" :address="address" :phone="phone" :email="email" :latlng="latlng" />
+      </sf-store-locator>
+      `
+    }),
+    {
+      info: {
+        summary: `
+          \`SfStore\` is a part of SfStoreLocator, and it allow to add stores to the map and the list at the same time.
+          Deeply configurable both with props and scss variables
+          <h2> Usage </h2>
+          <pre><code>import { SfStoreLocator } from "@storefrontui/vue"</code></pre>
           ${generateStorybookTable(scssTableConfig, "SCSS variables")}
           `
       }
@@ -213,7 +294,7 @@ storiesOf("Organisms|StoreLocator", module)
           default: object("stores", stores)
         }
       },
-      components: { SfStoreLocator, SfStore },
+      components: { SfStoreLocator },
       template: `
       <sf-store-locator :center="center" :zoom="zoom">
         <sf-store v-for="(store, index) in stores" :key="index" v-bind="store" />
@@ -229,7 +310,7 @@ storiesOf("Organisms|StoreLocator", module)
           \`SfStoreLocator\` allow to customise how the store icon is shown on the map by the use of the \`marker-icon\` slot.
           See the story source for an example
           <h2> Usage </h2>
-          <pre><code>import { SfStoreLocator, SfStore } from "@storefrontui/vue"</code></pre>
+          <pre><code>import { SfStoreLocator } from "@storefrontui/vue"</code></pre>
           ${generateStorybookTable(scssTableConfig, "SCSS variables")}
           `
       }
@@ -254,7 +335,7 @@ storiesOf("Organisms|StoreLocator", module)
           default: object("stores", stores)
         }
       },
-      components: { SfStoreLocator, SfStore },
+      components: { SfStoreLocator },
       template: `
       <sf-store-locator :center="center" :zoom="zoom">
         <sf-store v-for="(store, index) in stores" :key="index" v-bind="store" />
@@ -270,7 +351,7 @@ storiesOf("Organisms|StoreLocator", module)
           \`SfStoreLocator\` allow to customise the loading indicator by the use of the \`map-loading\` slot.
           See the story source for an example
           <h2> Usage </h2>
-          <pre><code>import { SfStoreLocator, SfStore } from "@storefrontui/vue"</code></pre>
+          <pre><code>import { SfStoreLocator } from "@storefrontui/vue"</code></pre>
           ${generateStorybookTable(scssTableConfig, "SCSS variables")}
           `
       }
