+++ packages/vue/src/components/organisms/SfStoreLocator/_internal/SfStore.vue
@@ -1,0 +1,2 @@
+<script src="./SfStore.js"></script>
+<template lang="html" src="./SfStore.html"></template>
